Add plugin which parses scene and gallery filenames on add
Add a plugin which tries to parse the names of scenes and galleries when
they're added / scanned. The primarily supported format is
`{studio}.{year}.{month}.{day}.{performer}.{performer}.{title}`. Studio
can optionally include ".com". Some "parts" are ignored like resolution,
some release group names etc. Performers can be separated by an optional
"and" which will be ignored as well. Performers are both searched based
on name and aliases, for which aliases are split based on `,` and `;`
and needs to result in an "exact" match meaning "First Last" won't match
a performer which name or alias is just "First". When there are multiple
matches (for example when filename is "first1.first2.last2" and multiple
performers having the "first1" alias) all matched performers will be
added.