Tons of changes, primarily related to Collections and Genres
Site and Studio Collections have been split and are user configurable, as well as being able to create Collections for Performers, specific Stash tags, and associated Movies

Also the user can opt to include images from any Galleries that are attached to the scenes.  Images can optionally be split into Posters and Background depending on orientation automatically

Performer tags in Stash can now be optionally added to the Scene Genres in Plex.  (So for example of an actress has the tag "Blonde", that can be auto added to all scenes with her in it)