Updating timestamp.trade and stash-performer-gallery for better background processing.
For timestamp.trade scene updates trigger a background task to do the processing so the scene update is not blocked.
A timestamp.trade url can be added to a scene and if this is present it will look up scene metadata for updates.
This allows you to re-process scenes for movie and gallery information after the initial task.

stashdb-performer-gallery has been updated to process in background jobs.