Added API Key option
Added API Key logic to allow FileMonitor to work in service mode when user has configured a Stash username and password.
To slim down the main config, moved self_unit_test and task_examples to their own configuration files.
Added more self_unit_test, and a selfUnitTest activation field that is normally false.
When field is true, it turns on the task for unit testing.