Parse All Galleries for date (#507)
* Parse All Galleries for date

- Break out the date parsing to a separate function.
- Expand gallery search to more than single file zip galleries.
- Also scan the gallery's folder path if available.

The final valid parsed date of the files, then the gallery folder will be taken as the gallery's date. This will preserve the previous behavior of single file zip galleries, as there will be only one path to ever consider.

* Revert to previous logic

Previous logic returned the last possible valid date string in the path string, which is the more correct approach when a parent directory might also have a date string.

* formatting

* bump version

* Cleanup lingering prototype line