Add TPDB to list of valid domains
This adds the TPDB domain to the list of valid stash boxes that can be detected by the "PerformerStashboxUrlToID" plugin. Because TPDB has multiple possible perfromer links, we need to add a check specific for links with the `/performers/` prefix that Stash-Box expects. This also adds a UUID check to make sure the Stash-Box ID is in the correct format, since TPDB is also a bit more relaxed about that.